darwin: clarify why we can't use poll
On latest MacOS versions, poll() behaves the same way as on Linux when
fds=NULL is passed. But we still can't use it because it doesn't work
with device files.

Since POSIX does not require poll() to do so, renamed the macro used to
conditionalize using poll() on darwin (and windows) to reflect the fact
it's not necessarily "broken", but only has undesired behavior.

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>